Move query-status.ts to /query-history
