+++ extensions/ql-vscode/src/query-history/history-tree-data-provider.ts
@@ -11,7 +11,7 @@ import {
 import { DisposableObject } from "../common/disposable-object";
 import { assertNever } from "../common/helpers-pure";
 import { QueryHistoryInfo } from "./query-history-info";
-import { QueryStatus } from "../query-status";
+import { QueryStatus } from "./query-status";
 import { HistoryItemLabelProvider } from "./history-item-label-provider";
 
 export enum SortOrder {
