+++ extensions/ql-vscode/src/query-history/query-status.ts
@@ -1,5 +1,5 @@
-import { assertNever } from "./common/helpers-pure";
-import { VariantAnalysisStatus } from "./variant-analysis/shared/variant-analysis";
+import { assertNever } from "../common/helpers-pure";
+import { VariantAnalysisStatus } from "../variant-analysis/shared/variant-analysis";
 
 export enum QueryStatus {
   InProgress = "InProgress",
