+++ extensions/ql-vscode/src/query-history/store/query-history-variant-analysis-domain-mapper.ts
@@ -23,7 +23,7 @@ import {
 } from "../../variant-analysis/shared/variant-analysis";
 import { assertNever } from "../../common/helpers-pure";
 import { QueryLanguage } from "../../common/query-language";
-import { QueryStatus } from "../../query-status";
+import { QueryStatus } from "../query-status";
 import { VariantAnalysisHistoryItem } from "../variant-analysis-history-item";
 
 export function mapQueryHistoryVariantAnalysisToDto(
