+++ extensions/ql-vscode/src/query-history/variant-analysis-history-item.ts
@@ -1,4 +1,4 @@
-import { QueryStatus } from "../query-status";
+import { QueryStatus } from "./query-status";
 import { VariantAnalysis } from "../variant-analysis/shared/variant-analysis";
 
 /**
