+++ extensions/ql-vscode/test/vscode-tests/no-workspace/query-history/query-history-info.test.ts
@@ -1,4 +1,4 @@
-import { QueryStatus } from "../../../../src/query-status";
+import { QueryStatus } from "../../../../src/query-history/query-status";
 import {
   buildRepoLabel,
   getActionsWorkflowRunUrl,
